Add support for reading the time when a dive started
Add the getDateTime method to the parser class. This method returns the
time when a dive started as a time_t. Note that this call has to make
assumptions about the time zone the dive computer was in at the time of
the dive, we currently assume that the computer parsing the data is in
the same time zone.